move DestinationAttribute to its own file
